+++ mettagrid/src/metta/mettagrid/util/hydra.py
@@ -9,14 +9,13 @@ from metta.common.util.fs import cd_repo_root
 
 
 def config_from_path(
-    config_path: str | DictConfig | ListConfig,
-    overrides: Optional[DictConfig | ListConfig | dict] = None,
+    config_path: str, overrides: Optional[DictConfig | ListConfig | dict] = None
 ) -> DictConfig | ListConfig:
     """
-    Load configuration from a path or return an already-loaded config.
+    Load configuration from a path, with better error handling
 
     Args:
-        config_path: Path to the configuration (string), or an already-loaded DictConfig/ListConfig
+        config_path: Path to the configuration
         overrides: Optional overrides to apply to the configuration
 
     Returns:
@@ -28,20 +27,15 @@ def config_from_path(
     if config_path is None:
         raise ValueError("Config path cannot be None")
 
-    # If config_path is already a config object, use it directly
-    if isinstance(config_path, (DictConfig, ListConfig)):
-        cfg = config_path
-    else:
-        # Original Hydra loading logic
-        cfg = hydra.compose(config_name=config_path)
+    cfg = hydra.compose(config_name=config_path)
 
-        # when hydra loads a config, it "prefixes" the keys with the path of the config file.
-        # We don't want that prefix, so we remove it.
-        if config_path.startswith("/"):
-            config_path = config_path[1:]
+    # when hydra loads a config, it "prefixes" the keys with the path of the config file.
+    # We don't want that prefix, so we remove it.
+    if config_path.startswith("/"):
+        config_path = config_path[1:]
 
-        for p in config_path.split("/")[:-1]:
-            cfg = cfg[p]
+    for p in config_path.split("/")[:-1]:
+        cfg = cfg[p]
 
     if overrides not in [None, {}]:
         # Allow overrides that are not in the config.
